% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Divergences_and_Distances.R
\name{BhattacharyyaDistance}
\alias{BhattacharyyaDistance}
\title{Calculates the bhattacharyya distance of the given probability distribution and the uniform distribution}
\usage{
BhattacharyyaDistance(prob)
}
\arguments{
\item{prob}{A numeric vector containing a probability distribution of ordinal patterns}
}
\value{
The Bhattacharya distance between the probability distribution of time series and uniform distribution
}
\description{
Calculates the bhattacharyya distance of the given probability distribution and the uniform distribution
}
\examples{
set.seed(123, kind = "Mersenne-Twister")
x <- runif(110000)
d <- 3
del <- 1
p <- BP(series = x, dimension = d, delay = del)
BhattacharyyaDistance(prob = p)

}
\references{
Kailath, Thomas. "The divergence and Bhattacharyya distance measures in signal selection." IEEE transactions on communication technology 15.1 (1967): 52-60.
}
\author{
Eduarda Chagas
}
