% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{BP}
\alias{BP}
\title{Calculates the Bandt-Pompe distribution of the given time series}
\usage{
BP(series, dimension, delay)
}
\arguments{
\item{series}{A numeric vector (e.g. a time series)}

\item{dimension}{Dimension size of ordinal patterns}

\item{delay}{Size of the embedding delay of ordinal patterns}
}
\value{
A numeric vector containing the probability distribution
}
\description{
Calculates the Bandt-Pompe distribution of the given time series
}
\examples{
set.seed(123, kind = "Mersenne-Twister")
x <- runif(110000)
d <- 3
del <- 1
BP(series = x, dimension = d, delay = del)

}
\references{
Bandt, Christoph, and Bernd Pompe. "Permutation entropy: a natural complexity measure for time series." Physical review letters 88.17 (2002): 174102.
}
\author{
Eduarda Chagas
}
