% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Planes.R
\name{.Cotas}
\alias{.Cotas}
\title{Plot curves of the time-causal entropy-complexity plane at a given dimension}
\usage{
.Cotas(dimension)
}
\description{
Plot curves of the time-causal entropy-complexity plane at a given dimension
}
\keyword{internal}
