% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{.DefineSymbols}
\alias{.DefineSymbols}
\title{Calculates all ordinal patterns with a given dimension}
\usage{
.DefineSymbols(dimension)
}
\description{
Calculates all ordinal patterns with a given dimension
}
\keyword{internal}
