% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entropies.R
\name{.ShannonEntropy}
\alias{.ShannonEntropy}
\title{Calculates the Shannon entropy of a probability distribution}
\usage{
.ShannonEntropy(prob)
}
\description{
Calculates the Shannon entropy of a probability distribution
}
\keyword{internal}
