% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Divergences_and_Distances.R
\name{.SquaredDistance}
\alias{.SquaredDistance}
\title{Calculates the squared distance of the given probability distribution and the uniform distribution}
\usage{
.SquaredDistance(prob)
}
\description{
Calculates the squared distance of the given probability distribution and the uniform distribution
}
\keyword{internal}
