% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{.Weights}
\alias{.Weights}
\title{Calculates the weights required in calculating the weighted permutation entropy}
\usage{
.Weights(series, dimension, delay)
}
\description{
Calculates the weights required in calculating the weighted permutation entropy
}
\keyword{internal}
