% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equalities_Values.R
\name{EqualitiesValues}
\alias{EqualitiesValues}
\title{Calculates the percentage of repeated elements in a time series}
\usage{
EqualitiesValues(series)
}
\arguments{
\item{series}{A numeric vector (e.g. a time series)}
}
\value{
Percentage of repeated elements of a time series
}
\description{
Calculates the percentage of repeated elements in a time series
}
\examples{
set.seed(123, kind = "Mersenne-Twister")
x <- runif(110000)
EqualitiesValues(series = x)
}
\author{
Eduarda Chagas
}
