% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Divergences_and_Distances.R
\name{EuclidianDistance}
\alias{EuclidianDistance}
\title{Calculates the euclidian distance of the given probability distribution and the uniform distribution}
\usage{
EuclidianDistance(prob)
}
\arguments{
\item{prob}{A numeric vector containing a probability distribution of ordinal patterns}
}
\value{
The euclidian distance between the probability distribution of time series and uniform distribution
}
\description{
Calculates the euclidian distance of the given probability distribution and the uniform distribution
}
\examples{
set.seed(123, kind = "Mersenne-Twister")
x <- runif(110000)
d <- 3
del <- 1
p <- BandtPompe(series = x, dimension = d, delay = del)
EuclidianDistance(prob = p)
}
\author{
Eduarda Chagas
}
