% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entropies.R
\name{FisherEntropy}
\alias{FisherEntropy}
\title{Calculates the Fisher entropy of a probability distribution}
\usage{
FisherEntropy(prob)
}
\arguments{
\item{prob}{A numeric vector containing a probability distribution of ordinal patterns}
}
\value{
The entropy of Fisher
}
\description{
Calculates the Fisher entropy of a probability distribution
}
\examples{
set.seed(123, kind = "Mersenne-Twister")
x <- runif(110000)
d <- 3
del <- 1
p <- BandtPompe(series = x, dimension = d, delay = del)
FisherEntropy(prob = p)
}
\references{
Olivares, Felipe, Angelo Plastino, and Osvaldo A. Rosso. "Contrasting chaos with noise via local versus global information quantifiers." Physics Letters A 376.19 (2012): 1577-1583.
}
\author{
Eduarda Chagas
}
