% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Planes.R
\name{FSPlane}
\alias{FSPlane}
\title{Plots the Fisher-Shannon Plane for a given probability}
\usage{
FSPlane(prob)
}
\arguments{
\item{prob}{A numeric vector containing a probability distribution of ordinal patterns}
}
\value{
A Plot with FS Plane
}
\description{
Plots the Fisher-Shannon Plane for a given probability
}
\examples{
set.seed(123, kind = "Mersenne-Twister")
x1 <- runif(10000)
x2 <- runif(10000)
x3 <- runif(10000)
d <- 3
del <- 1
probability <- matrix(nrow = 3, ncol = factorial(d))
probability[1,] <- BandtPompe(series = x1, dimension = d, delay = del)
probability[2,] <- BandtPompe(series = x2, dimension = d, delay = del)
probability[3,] <- BandtPompe(series = x3, dimension = d, delay = del)
p <- FSPlane(prob = probability)
print(p)

}
\references{
Rosso, Osvaldo A., Felipe Olivares, and Angelo Plastino. "Noise versus chaos in a causal Fisher-Shannon plane." arXiv preprint arXiv:1510.06950 (2015).
}
\author{
Eduarda Chagas
}
