% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Divergences_and_Distances.R
\name{KullbackDivergence}
\alias{KullbackDivergence}
\title{Calculates the kullback divergence of the given probability distribution and the uniform distribution}
\usage{
KullbackDivergence(prob)
}
\arguments{
\item{prob}{A numeric vector containing a probability distribution of ordinal patterns}
}
\value{
The Kullback divergence between the probability distribution of time series and uniform distribution
}
\description{
Calculates the kullback divergence of the given probability distribution and the uniform distribution
}
\examples{
set.seed(123, kind = "Mersenne-Twister")
x <- runif(110000)
d <- 3
del <- 1
p <- BandtPompe(series = x, dimension = d, delay = del)
KullbackDivergence(prob = p)
}
\references{
Van Erven, Tim, and Peter Harremos. "Rényi divergence and Kullback-Leibler divergence." IEEE Transactions on Information Theory 60.7 (2014): 3797-3820.
}
\author{
Eduarda Chagas
}
