% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{OPTG}
\alias{OPTG}
\title{Calculates the transition graph of ordinal pattern for the given time series}
\usage{
OPTG(series, dimension, delay)
}
\arguments{
\item{series}{A numeric vector (e.g. a time series)}

\item{dimension}{Dimension size of ordinal patterns}

\item{delay}{Size of the embedding delay of ordinal patterns}
}
\value{
A numeric matrix containing the graph
}
\description{
Calculates the transition graph of ordinal pattern for the given time series
}
\examples{
set.seed(123, kind = "Mersenne-Twister")
x <- runif(110000)
d <- 3
del <- 1
OPTG(series = x, dimension = d, delay = del)

}
\references{
Borges, João B., et al. "Learning and distinguishing time series dynamics via ordinal patterns transition graphs." Applied Mathematics and Computation 362 (2019): 124554.
}
\author{
Eduarda Chagas
}
