% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Planes.R
\name{PlotSeries}
\alias{PlotSeries}
\title{Plot time series}
\usage{
PlotSeries(series)
}
\arguments{
\item{series}{A numeric vector (e.g. a time series)}
}
\value{
A Time series plot
}
\description{
Plot time series
}
\examples{
set.seed(123, kind = "Mersenne-Twister")
x <- runif(110000)
p <- PlotSeries(series = x)
print(p)

}
\author{
Eduarda Chagas
}
