% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entropies.R
\name{RenyiEntropy}
\alias{RenyiEntropy}
\title{Calculates the Renyi entropy of a probability distribution}
\usage{
RenyiEntropy(prob, q)
}
\arguments{
\item{prob}{A numeric vector containing a probability distribution of ordinal patterns}

\item{q}{The entropy order. It allows only positive numbers.}
}
\value{
The entropy of Renyi
}
\description{
Calculates the Renyi entropy of a probability distribution
}
\examples{
set.seed(123, kind = "Mersenne-Twister")
x <- runif(110000)
d <- 3
del <- 1
p <- BandtPompe(series = x, dimension = d, delay = del)
RenyiEntropy(prob = p, q = 0.01)
}
\references{
Lenzi, E. K., R. S. Mendes, and L. R. Da Silva. "Statistical mechanics based on Renyi entropy." Physica A: Statistical Mechanics and its Applications 280.3-4 (2000): 337-345.
}
\author{
Eduarda Chagas
}
