% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entropies.R
\name{ShannonEntropy}
\alias{ShannonEntropy}
\title{Calculates the normalized Shannon entropy of a probability distribution}
\usage{
ShannonEntropy(prob)
}
\arguments{
\item{prob}{A numeric vector containing a probability distribution of ordinal patterns}
}
\value{
The entropy of Shannon Normalized
}
\description{
Calculates the normalized Shannon entropy of a probability distribution
}
\examples{
set.seed(123, kind = "Mersenne-Twister")
x <- runif(110000)
d <- 3
del <- 1
p <- BP(series = x, dimension = d, delay = del)
ShannonEntropy(prob = p)

}
\references{
Bandt, Christoph, and Bernd Pompe. "Permutation entropy: a natural complexity measure for time series." Physical review letters 88.17 (2002): 174102.
}
\author{
Eduarda Chagas
}
