% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entropies.R
\name{TsallisEntropy}
\alias{TsallisEntropy}
\title{Calculates the Tsallis entropy of a probability distribution}
\usage{
TsallisEntropy(prob, q)
}
\arguments{
\item{prob}{A numeric vector containing a probability distribution of ordinal patterns}

\item{q}{The entropy order. It allows only positive numbers.}
}
\value{
The entropy of Tsallis
}
\description{
Calculates the Tsallis entropy of a probability distribution
}
\examples{
set.seed(123, kind = "Mersenne-Twister")
x <- runif(110000)
d <- 3
del <- 1
p <- BandtPompe(series = x, dimension = d, delay = del)
TsallisEntropy(prob = p, q = 0.01)
}
\references{
De Albuquerque, M. Portes, Israel A. Esquef, and AR Gesualdi Mello. "Image thresholding using Tsallis entropy." Pattern Recognition Letters 25.9 (2004): 1059-1065.
}
\author{
Eduarda Chagas
}
