% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{WATG}
\alias{WATG}
\title{Calculates the weighted transition graph of ordinal pattern for the given time series}
\usage{
WATG(series, dimension, delay)
}
\arguments{
\item{series}{A numeric vector (e.g. a time series)}

\item{dimension}{Dimension size of ordinal patterns}

\item{delay}{Size of the embedding delay of ordinal patterns}
}
\value{
A numeric matrix containing the graph
}
\description{
Calculates the weighted transition graph of ordinal pattern for the given time series
}
\examples{
set.seed(123, kind = "Mersenne-Twister")
x <- runif(110000)
d <- 3
del <- 1
WATG(series = x, dimension = d, delay = del)

}
\references{
Chagas, Eduarda T. C., et al. "Analysis and Classification of SAR Textures using Information Theory" (2020)
}
\author{
Eduarda Chagas
}
