% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Divergences_and_Distances.R
\name{WoottersDistance}
\alias{WoottersDistance}
\title{Calculates the wootters distance of the given probability distribution and the uniform distribution}
\usage{
WoottersDistance(prob)
}
\arguments{
\item{prob}{A numeric vector containing a probability distribution of ordinal patterns}
}
\value{
The Wootters distance between the probability distribution of time series and uniform distribution
}
\description{
Calculates the wootters distance of the given probability distribution and the uniform distribution
}
\examples{
set.seed(123, kind = "Mersenne-Twister")
x <- runif(110000)
d <- 3
del <- 1
p <- BandtPompe(series = x, dimension = d, delay = del)
WoottersDistance(prob = p)
}
\references{
Majtey, A., et al. "Wootters’ distance revisited: a new distinguishability criterium." The European Physical Journal D-Atomic, Molecular, Optical and Plasma Physics 32.3 (2005): 413-419.
}
\author{
Eduarda Chagas
}
