% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{WPE}
\alias{WPE}
\title{Calculates the Weighted Bandt-Pompe distribution of the given time series}
\usage{
WPE(series, dimension, delay)
}
\arguments{
\item{series}{A numeric vector (e.g. a time series)}

\item{dimension}{Dimension size of ordinal patterns}

\item{delay}{Size of the embedding delay of ordinal patterns}
}
\value{
A numeric vector containing the probability distribution
}
\description{
Calculates the Weighted Bandt-Pompe distribution of the given time series
}
\examples{
set.seed(123, kind = "Mersenne-Twister")
x <- runif(110000)
d <- 3
del <- 1
WPE(series = x, dimension = d, delay = del)
}
\references{
Fadlallah, Bilal, et al. "Weighted-permutation entropy: A complexity measure for time series incorporating amplitude information." Physical Review E 87.2 (2013): 022911.
}
\author{
Eduarda Chagas
}
